requirements.txt
Django==1.7.1
Pillow==2.6.1
django-bootstrap-form==3.1
django-debug-toolbar==1.2.2
django-registration-redux==1.1
django-widget-tweaks==1.3
flake8==2.2.5
ipdb==0.8
ipython==2.3.0
markdown2==2.3.0
mccabe==0.2.1
pep8==1.5.7
pyflakes==0.8.1
python-dateutil==2.2
pytz==2014.9
six==1.8.0
sqlparse==0.1.13
wsgiref==0.1.2
